gdeploy/requirements.txt
requests == 2.21.0
numpy==1.16.1
google-cloud-bigquery==1.18.0
google-cloud-storage==1.18.0
google-cloud-core==1.0.3
opencv-python==4.1.1.26
Pillow == 6.0.0
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp37-cp37m-linux_x86_64.whl